05/09/05 -- CJH -- Multidrizzle 2.6.4 check-in.  Modified the fix provided in Multidrizzle 2.6.3 to delete any association tables created to address the zero exposure time problem upon completion of Multidrizzle processing.
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@360 fe389314-cf27-0410-b35b-8c050e845b92

+++ multidrizzle/lib/__init__.py
@@ -1548,9 +1548,10 @@ help file.
                 # Now have image_manager remove all image products
                 self.image_manager.removeMDrizProducts()
                 
-            # Remove the self.zeroExptimeAsnTable file if it exisits
-            if os.path.exists(self.zeroExptimeAsnTable):
-                os.remove(self.zeroExptimeAsnTable)
+            # Remove the self.zeroExptimeAsnTable file if it exists
+            if self.zeroExptimeAsnTable != None:
+                if os.path.exists(self.zeroExptimeAsnTable):
+                    os.remove(self.zeroExptimeAsnTable)
 
         if self.pars.switches['timing']:
             self.steps.reportTimes()
